164 some fps values throws off syncing update with fixed update

+++ UniTAS/Plugin/Patches/TimePatch.cs
@@ -306,7 +306,7 @@ public class TimePatch
 
         private static bool Prefix(ref float __result)
         {
-            __result = TimeEnv.FrameTime;
+            __result = (float)TimeEnv.FrameTime;
             return false;
         }
     }
